move TabularPercentLegend into NivoCharts
